pep8 and start to work on the OpenCL buffer allocation
the idea is to factorize the buffer allocation and the memory check
for all OpenCL mem buffers.
so instead of writing a dedicated code for the allocation, I fill a simple
list (buffers) which contain all the information necessary to do the job
(name, flag, dtype , size)

this way it will be possible later to create a generic functions which do the allocation.