Provide a script for sparsification of a single crystal datset:
Allows the (lossy) compression of single crystal datasets.
By default, the compression ratio is about 100x,

The background is recorded assuming a polar symmetry.
The signal is sigma-clipped for 5 times (default threshold: 5 sigmas)
Then peaks are recorded (default threshold: 3 sigma).

Data are packed in HDF5 and provide a script to re-densify the data.

Close #1405