API: restore back compatibility in histoBBox1d and histoBBox2d
The signatures of a number of functions were changed in #1528.  This restores
the ability to pass by keyword pos0Range and pos1Range in histoBBox1d and
histoBBox2D.

If both the old and new parameters are passed, `ValueError` is raised, if only
the old keyword is passed a warning is issued.  If the value is passed
as a positional the users is unaware of this change.